Add more examples of clip and vit for training/inference (#12)
* refactor: remove dtype argument from sharded_init initializers

* chore: add empty vit_training.py example file

* feat: add tensorflow and tensorflow-datasets to dev dependencies in pixi

* feat: add concise ViT training loop for MNIST using tfds and jimm

* feat: update ViT model config and improve dataset pipeline in training script

* fix: add drop_remainder=True to batch to prevent assertion error on last batch

* feat: add validation and test evaluation to ViT MNIST training script

* feat: add CLIP inference example using JAX and Flax

* feat: add Jupyter notebook to display the architecture of the models

+++ src/jimm/models/clip.py
@@ -56,10 +56,10 @@ class VisionTransformer(nnx.Module):
             bias_init=sharded_init(nnx.initializers.zeros_init(), P("model"), mesh),
         )
         _cls_token_initializer = sharded_init(nnx.initializers.zeros_init(), P(None, None, "model"), mesh)
-        cls_token_value: Float[Array, "1 1 width"] = _cls_token_initializer(rngs.params(), (1, 1, width), dtype=dtype)
+        cls_token_value: Float[Array, "1 1 width"] = _cls_token_initializer(rngs.params(), (1, 1, width))
         self.cls_token = nnx.Param(cls_token_value)
         _position_embeddings_initializer = sharded_init(nnx.initializers.truncated_normal(stddev=0.02), P(None, None, "model"), mesh)
-        pos_emb_value: Float[Array, "1 n_patches+1 width"] = _position_embeddings_initializer(rngs.params(), (1, n_patches + 1, width), dtype=dtype)
+        pos_emb_value: Float[Array, "1 n_patches+1 width"] = _position_embeddings_initializer(rngs.params(), (1, n_patches + 1, width))
         self.position_embeddings = nnx.Param(pos_emb_value)
 
         self.ln_pre = nnx.LayerNorm(
@@ -213,7 +213,7 @@ class CLIP(nnx.Module):
             rngs=rngs,
             embedding_init=sharded_init(nnx.initializers.xavier_uniform(), P("model", None), mesh),
         )
-        self.positional_embedding = nnx.Param(sharded_init(nnx.initializers.truncated_normal(stddev=0.02), P("model", None), mesh)(rngs.params(), (context_length, transformer_width), dtype=dtype))
+        self.positional_embedding = nnx.Param(sharded_init(nnx.initializers.truncated_normal(stddev=0.02), P("model", None), mesh)(rngs.params(), (context_length, transformer_width)))
         self.ln_final = nnx.LayerNorm(
             transformer_width,
             epsilon=1e-5,
@@ -232,7 +232,7 @@ class CLIP(nnx.Module):
             rngs=rngs,
             kernel_init=sharded_init(nnx.initializers.xavier_uniform(), P("model", None), mesh),
         )
-        self.logit_scale = nnx.Param(sharded_init(nnx.initializers.ones_init(), P("model"), mesh)(rngs.params(), (), dtype=dtype))
+        self.logit_scale = nnx.Param(sharded_init(nnx.initializers.ones_init(), P("model"), mesh)(rngs.params(), ()))
 
     def encode_image(self, image: Float[Array, "batch height width channels"]) -> Float[Array, "batch transformer_width"]:
         """
